feat: Add styling options for frame creation
