+++ asciic/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-clap = "3.2.22"
+clap = { version = "3.2.22", features = ["derive"] }
 ctrlc = { version = "3.2.3", features = ["termination"] }
 image = "0.24.4"
 rayon = "1.5.3"
